+++ lib/ChangeLog.md
@@ -1,5 +1,10 @@
 # Revision history for irc-core
 
+## 2.1.1.0  -- 2016-08-13
+
+* Add `Eq` instances to `UserInfo` and `RawIrcMsg`
+* Add `IsString` instance to `Identifier`
+
 ## 2.1.0.0  -- 2016-08-13
 
 * Add BatchStart and BatchEnd messages